services/etl/src/requirements.txt
elasticsearch==8.17.2
psycopg==3.2.4
psycopg-binary==3.2.4
python-dotenv==1.0.1
pydantic-settings==2.8.1
redis==5.2.1
